login y cambios en form de datos calculo

+++ application/controllers/Administrador.php
@@ -2,63 +2,53 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Administrador extends CI_Controller {
-
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/index.php/welcome
-	 *	- or -
-	 * 		http://example.com/index.php/welcome/index
-	 *	- or -
-	 * Since this controller is set as the default controller in
-	 * config/routes.php, it's displayed at http://example.com/
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /index.php/welcome/<method_name>
-	 * @see http://codeigniter.com/user_guide/general/urls.html
-	 */
+	
 
 	public function __construct()
 	{
 		parent::__construct();
 		$this->load->model("Administrador_model");
-        //$this->load->library('session');
-
-
+		$this->load->library('session');
 		$this->load->helper('url_helper');
 		$this->load->helper('vayes_helper');
 
 
 	}
 	public function index()
-	{
-		$this->load->view('crud/header');	
-		$data['datos'] = $this->Administrador_model->getData();	
-		$this->load->view('crud/Administracion',$data);		
-		$this->load->view('crud/footer');
+	{			
+		if($this->session->userdata())
+		{	
+			$this->load->view('crud/header');	
+			$data['datos'] = $this->Administrador_model->getData();	
+			//$this->load->view('crud/menu');
+			$this->load->view('crud/Administracion',$data);		
+			$this->load->view('crud/footer');			
+		}
+		else{
+			redirect(base_url() . 'Login');		
+		}
+
 	}
 
-	public function nuevo($id=null)
-	{
-		
-	}
+
 
 	public function create()
 	{
-		$datos = $this->input->post();
-		$precio = $this->input->post('valor');
 
-		if(isset($datos))
-		{
+		if($this->session->userdata())
+		{	
+
+			$datos = $this->input->post();
+			$precio = $this->input->post('valor');
+
 			$montos_cuota = $this->calcula_cuota_mes($precio, 0.0045, 300);
 			//var_dump($montos_cuota);
 			//exit;
 
-			
-				$a=$montos_cuota['cuota'];
-				$b=$montos_cuota['sueldo_ideal'];				
-		
+
+			$a=$montos_cuota['cuota'];
+			$b=$montos_cuota['sueldo_ideal'];				
+
 
 			$data = array(            
             'descripcion' => $this->input->post('descripcion'), //input
@@ -72,23 +62,26 @@ class Administrador extends CI_Controller {
 			$this->db->insert('condominio', $data);
 			redirect(base_url() . 'Administrador');
 		}
-		else {
-			redirect(base_url());
+		else{
+			redirect(base_url() . 'Login');		
 		}
+
+
+
+
 	}
 	public function update()
 	{
-		$datos = $this->input->post();
-		if(isset($datos))
-		{
+
+		if($this->session->userdata())
+		{	
+			$datos = $this->input->post();		
 			$precio = $this->input->post('valor_e');
 			$montos_cuota = $this->calcula_cuota_mes($precio, 0.0045, 300);
 			//var_dump($montos_cuota);
-			//exit;
-
-			
-				$a=$montos_cuota['cuota'];
-				$b=$montos_cuota['sueldo_ideal'];	
+			//exit;			
+			$a=$montos_cuota['cuota'];
+			$b=$montos_cuota['sueldo_ideal'];	
 			$id = $this->input->post('id_e'); //input
 			$data = array(            
             'descripcion' => $this->input->post('descripcion_e'), //input
@@ -102,19 +95,25 @@ class Administrador extends CI_Controller {
 
 			$this->db->where('id', $id);
 			$this->db->update('condominio', $data);
-
 			redirect(base_url() . 'Administrador');
+		}
+		else{
+			redirect(base_url() . 'Login');	
+		}
 
-		}
-		else {
-			redirect(base_url());
-		}
+
 	}
 
 	public function delete($id=null)
 	{
-		$this->db->delete('condominio', array('id' => $id));
-		redirect(base_url() . 'Administrador');
+		if($this->session->userdata())
+		{	
+			$this->db->delete('condominio', array('id' => $id));
+			redirect(base_url() . 'Administrador');
+		}
+		else{
+			redirect(base_url() . 'Login');	
+		}
 	}
 	function calcula_cuota_mes($monto_prestamo = null, $porcentaje = null, $cuotas = null)
 	{
